서버 라우터 구조 개선
