+++ sunghyun/.idea/workspace.xml
@@ -5,17 +5,24 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="0019cf39-7305-4b03-ae98-aff8bf4bcdea" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/server/dbConnection.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/server/router/login.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/server/router/todo.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/client/src/TodoListPage.js" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/TodoListPage.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/client/src/components/LoginInput.js" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/components/LoginInput.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/client/src/components/SignUpInput.js" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/components/SignUpInput.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/client/src/index.js" beforeDir="false" afterPath="$PROJECT_DIR$/client/src/index.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/server/index.js" beforeDir="false" afterPath="$PROJECT_DIR$/server/index.js" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
     <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="FileTemplateManagerImpl">
+    <option name="RECENT_TEMPLATES">
+      <list>
+        <option value="HTML File" />
+        <option value="JavaScript File" />
+      </list>
+    </option>
   </component>
   <component name="Git.Settings">
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$/.." />
@@ -30,6 +37,7 @@
     <option name="showLibraryContents" value="true" />
   </component>
   <component name="PropertiesComponent">
+    <property name="DefaultHtmlFileTemplate" value="HTML File" />
     <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
     <property name="RunOnceActivity.ShowReadmeOnStart" value="true" />
     <property name="WebServerToolWindowFactoryState" value="false" />
@@ -53,11 +61,22 @@
       <workItem from="1641869113662" duration="721000" />
       <workItem from="1641969741218" duration="8660000" />
       <workItem from="1642234894408" duration="336000" />
-      <workItem from="1642515322556" duration="937000" />
+      <workItem from="1642515322556" duration="2848000" />
     </task>
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
     <option name="version" value="3" />
+  </component>
+  <component name="Vcs.Log.Tabs.Properties">
+    <option name="TAB_STATES">
+      <map>
+        <entry key="MAIN">
+          <value>
+            <State />
+          </value>
+        </entry>
+      </map>
+    </option>
   </component>
 </project>
